Bug:  Conversion of a Clover LONG data type into the databases REAL data type fails since it attempts to perform a numeric copy and LONG isn't a numeric.
Test Case:  Create a database of type NUMBER(38, 0), actually bigger than a long which is about 19 digits.  Create a data record that is a LONG data type.  Attempt to store the data record into the database.  Get an exception.

Fix:  If the Clover data type is Long then perform a CopyLong instead of CopyNumeric.


git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@530 a09ad3ba-1a0f-0410-b1b9-c67202f10d70